refactor: extract shared storage path helpers into storage.rs
Move create_id, relative_path_for, and resolve_stored_path into a
shared storage module, eliminating duplication between icon_handlers
and font_handlers.
